requirements.txt
# Core dependencies
google-genai>=0.2.0
python-dotenv>=1.0.0
httpx>=0.27.0
rich>=13.7.0

# Contextual AI (if available as package)
# contextual-client>=0.1.0

# Optional: API server dependencies (not used in CLI)
# fastapi>=0.115.0
# uvicorn>=0.32.0
# python-multipart>=0.0.9
# sse-starlette>=2.0.0

# Development dependencies
# pytest>=8.0.0
# pytest-asyncio>=0.24.0